+++ apps/zhi-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-cli
 
+## 1.5.10
+
+### Patch Changes
+
+- zhi-log@1.14.10
+
+## 1.5.9
+
+### Patch Changes
+
+- zhi-log@1.14.9
+
 ## 1.5.8
 
 ### Patch Changes