+++ libs/zhi-blog-api/CHANGELOG.md
@@ -1,5 +1,34 @@
 # zhi-blog-api
 
+## 1.6.10
+
+### Patch Changes
+
+- fix deps
+- Updated dependencies
+  - zhi-lib-base@0.4.1
+
+## 1.6.9
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-lib-base@0.4.0
+
+## 1.6.8
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-lib-base@0.3.1
+
+## 1.6.7
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-lib-base@0.3.0
+
 ## 1.6.6
 
 ### Patch Changes