+++ libs/zhi-device/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zhi-device
 
+## 2.3.0
+
+### Minor Changes
+
+- add custom time for reload
+
+## 2.2.0
+
+### Minor Changes
+
+- make a better check for widget
+
 ## 2.1.0
 
 ### Minor Changes