+++ libs/zhi-siyuan-api/CHANGELOG.md
@@ -1,5 +1,56 @@
 # zhi-siyuan-api
 
+## 1.9.7
+
+### Patch Changes
+
+- fix deps
+- Updated dependencies
+  - zhi-blog-api@1.6.10
+  - zhi-lib-base@0.4.1
+  - zhi-common@1.11.2
+
+## 1.9.6
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-lib-base@0.4.0
+  - zhi-blog-api@1.6.9
+  - zhi-common@1.11.1
+
+## 1.9.5
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.11.0
+
+## 1.9.4
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-common@1.10.0
+
+## 1.9.3
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-lib-base@0.3.1
+  - zhi-blog-api@1.6.8
+  - zhi-common@1.9.8
+
+## 1.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - zhi-lib-base@0.3.0
+  - zhi-blog-api@1.6.7
+  - zhi-common@1.9.7
+
 ## 1.9.1
 
 ### Patch Changes